Drop the 'fparent' concept for 'noreply'
I was getting tangled in the fparent vs parent issue when looking up vars. If a node has a parent, it's in @node['parent']. If it should not reply to it when done let's set @node['noreply'] to true. That's all.

The var lookup code got simpler and the only big change is the
```ruby
  def wrap(h={})
    # ...
    m['nid'] = @node['noreply'] ? nil : parent
    # ...
```
check in #wrap.

This 'noreply' flag can be reused for concepts other that flanking.