Split Storage#consume in two methods
Split into  #consume_and_archive and #consume_and_discard.
Use a transaction [explicitely] only in #consume_and_archive since it has two database calls.

(I think that the delete in #consume_and_discard is also wrapped in a transaction, but I have the impression it is because of the Sequel sqlite3 adapter implementation)

gh-14