Allow for `r/egex/` instead of `(/egex/)`
When a regular expression is used as an argument in a line, most of the time, the parser mistakes that for a division.

By prefixing the regex first slash with a "r" (or "regex", "rex", "re", "rx"), the parser knows it's a regular expression coming, no need to wrap in parentheses.